+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this repository will be documented in this file.
 
 ## [Unreleased] - 2022-xx-xx
 
+### Added
+- Hyperparameter search for internal training jobs
+
 ### Fixed
 - Fix predict dataset for MQA model